Fix: Corrige error de autenticación en funciones de Firebase
Se actualizó la firma de las funciones `placeBet_crash` y `cashOut_crash` para que utilicen `request.auth` en lugar de `context.auth`, que estaba obsoleto en la v2 del SDK de Firebase Functions. Esto soluciona el error 401 (No autorizado) que se producía al intentar realizar una apuesta.

Este cambio es una versión simplificada y enfocada de un intento anterior, y se ha verificado que no introduce errores de compilación.
