+++ functions/.eslintrc.js
@@ -1,0 +1,19 @@
+module.exports = {
+  root: true,
+  env: {
+    es6: true,
+    node: true,
+  },
+  extends: [
+    "eslint:recommended",
+    "google",
+  ],
+  parserOptions: {
+    "ecmaVersion": 2020,
+  },
+  rules: {
+    "quotes": ["error", "double"],
+    "require-jsdoc": 0,
+    "no-unused-vars": ["warn", { "argsIgnorePattern": "^_" }],
+  },
+};
